refactor(worker): share memory content preparation across scopes
